refactor : split auth presentation model to request and response
